Can’t get the partition hash right?

+++ shared/Utils/Util.h
@@ -10,7 +10,7 @@ public:
     static QString HashPassword(QString username, QString password);
     static QString GenerateToken(QString username);
 
-    static qint16 getIntFromTwoInt(qint16 a, qint16 b)
+    static qint32 getIntFromTwoInt(qint32 a, qint32 b)
     {
         return a << 16 | (b & 0xFFFF);
     }
