+++ worldserver/Game/Maps/Map.h
@@ -6,18 +6,18 @@
 #include "Entities/ObjectMgr.h"
 #include "Partition.h"
 
-typedef QHash<qint16/*partitionId*/, Partition*> PartitionHash;
+typedef QHash<qint32/*partitionId*/, Partition*> PartitionHash;
 
 class Map
 {
 public:
-    Map(qint16 id);
+    Map(qint32 id);
     ~Map();
 
     qint16 GetId() { return m_id; }
 
-    Partition* CreatePartition(qint16 id);
-    Partition* FindPartition(qint16 id) const
+    Partition* CreatePartition(qint32 id);
+    Partition* FindPartition(qint32 id) const
     {
         PartitionHash::const_iterator it = m_partitions.find(id);
         return (it != m_partitions.end() ? it.value() : nullptr);
@@ -26,7 +26,7 @@ public:
     void AddCharacterToMap(Character* character);
 
 private:
-    qint16 m_id;
+    qint32 m_id;
     PartitionHash m_partitions;
 };
